requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
alphashape==1.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
beautifulsoup4==4.14.2
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
click-log==0.4.0
Crawl4AI==0.7.7
cryptography==46.0.3
cssselect==1.3.0
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fake-http-header==0.3.5
fake-useragent==2.2.0
fastapi==0.122.0
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.2
fastar==0.8.0
fastuuid==0.14.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
greenlet==3.2.4
grpcio==1.67.1
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.1.6
humanize==4.14.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
itemadapter==0.12.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
json5==0.12.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lark==1.3.1
litellm==1.80.7
lxml==5.4.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
networkx==3.6
nltk==3.9.2
numpy==2.3.5
openai==2.8.1
packaging==25.0
patchright==1.56.0
pillow==12.0.0
playwright==1.56.0
propcache==0.4.1
psutil==7.1.3
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
pyOpenSSL==25.3.0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
rank-bm25==0.2.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
rpds-py==0.29.0
rtree==1.4.1
scipy==1.16.3
scrapeops_python_logger==0.4.7
scrapeops_python_requests==0.4.7
sentry-sdk==2.46.0
shapely==2.1.2
shellingham==1.5.4
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.8
starlette==0.50.0
tf-playwright-stealth==1.2.0
tiktoken==0.12.0
tld==0.13.1
tokenizers==0.22.1
tqdm==4.67.1
trimesh==4.10.0
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
